third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/css-live-edit-expected.txt

Tests that styles are updated when live-editing css resource.


Running: testLiveEdit
display: block;
    block - div user agent stylesheet
font-size: 20px;
    20px - body css-live-edit.css:1 -> css-live-edit.css:1:7
[expanded] 
element.style { ()
